Add Aggregation class.
