--- ThingIFSDK/ThingIFSDK/FieldType.swift
@@ -1,29 +1,0 @@
-//
-//  Field Type.swift
-//  ThingIFSDK
-//
-//  Created on 2017/01/11.
-//  Copyright (c) 2017 Kii. All rights reserved.
-//
-
-import Foundation
-
-/** Field types to count. */
-public enum FieldType: String {
-
-    /** A type of a int field. */
-    case integer = "INTEGER"
-
-    /** A type of a decimal field. */
-    case decimal = "DECIMAL"
-
-    /** A type of a boolean field. */
-    case bool = "BOOL"
-
-    /** A type of a object field. */
-    case object = "OBJECT"
-
-    /** A type of a array field. */
-    case array = "ARRAY"
-}
-
